ai_test_app/requirements.txt
fastapi
uvicorn[standard]
jinja2
reportlab
openpyxl
google-generativeai
python-dotenv